pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "gendiff"
readme = "README.md"
authors = [{name = "AlexNeishchenko", email = "br.cheg@mail.ru"}]
requires-python = ">=3.8.1"
dependencies = [
    "pytest>=8.3.5",
    "pyyaml>=6.0.2",
    "pytest-cov >= 4.0.0",
    "ruff>=0.9.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["gendiff"]


[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"
